Recognize that variable vars are not assignments (#88)
* Tests: Add tests for variable variables

* Recognize that variable vars are not assignments

When a variable variable is used as the left side of an assignment,
there are actually two variables being checked: there is the current
one, and the variable one whose name is the value of the current one.

In this case, we must not treat the current one as an assignment; it is
actually a read. This change updates the logic so that this is the
case.