Composer: move PHPCS dependency to `require`
PHPCS is a direct dependency for this utility to be run, so should be in `require`.

The problem with not having it there is that other external /project rulesets may have different PHPCS requirements and Composer will not be able to take the requirements for this project into consideration when doing the version negotiations unless the dependency is explicitly set in `require`.

Relying on PHPCS being installed with the DealerDirect plugin is not stable as they may change their minimum required PHPCS version without notice.

As the minimum version in the `composer.json` file and the one in the `README` were not in line (though very close), I've also updated the minimum required version as mentioned in the `README` to be in line.

Question: as the Composer install instructions already suggest using the DealerDirect plugin, why not make that a `no-dev` requirement as well ?