Bug fix: $this in nested function declaration
While `$this` can be used freely within classes, traits and closures, if a nested closed scope is defined within any of these, `$this` is not inherited and should be considered undefined.

As things were, the sniff would not throw a warning for this (false negative).
This commit fixes this.

Includes unit tests.

Includes removing an outdated comment. `T_CLOSURE` works perfectly fine with the conditions array (and has for quite a few years).

Fixes 109