PHPStan config: don't report unmatched errors
Now the `PHP_CodeSniffer\Util\Tokens` file is imported, the `Constant T_\w+ not found` errors should no longer be thrown.

However, it depends on the order in which the files are analysed as that file is not (yet) `use`d in all files using the `T_` constants.

If I remove the `ignoreErrors`, [the build fails](https://circleci.com/gh/sirbrillig/phpcs-variable-analysis/281) on those errors for the `VariableAnalysisSniff` file, while if I leave the exclusion in, [the build fails](https://circleci.com/gh/sirbrillig/phpcs-variable-analysis/283) on a "Ignored error not matched".

So, for now, I'm leaving the ignore pattern in place and turning `reportUnmatchedIgnoredErrors` off to allow the build to pass.