Allow running the unit tests on more recent PHPUnit versions
I've tested and the unit tests as-is are compatible with PHPUnit 7.x and 8.x without needing any adjustments, so we may as well allow using/installing those versions.

Includes widening the version requirements for the `limedeck/phpunit-detailed-printer` to prevent it from being a blocker for using more recent PHPUnit versions.

Version 4.x of the `limedeck/phpunit-detailed-printer` is compatible with PHPUnit 7.x, version 5.x is compatible with PHPUnit 8.x.
Ref: https://github.com/LimeDeck/phpunit-detailed-printer/releases