Composer: remove the Composer PHPCS plugin
The Dealerdirect Composer PHPCS plugin is already a (non-dev) requirement of PHPCSUtils, so will be installed whether you require it or not.

By removing it from the `composer.json` file, potential conflicts between the versions allowed by PHPCSUtils and this plugin are prevented.

Note: The plugin has released a new version last week and the new version of PHPCSUtils which was also released a few days back, already allows for it. The new version of the plugin includes support for Composer 2.0 and allows for testing standards with the plugin on PHP 8 and with PHPCS 4.x-dev.

Ref: https://github.com/Dealerdirect/phpcodesniffer-composer-installer/releases/tag/v0.7.0