Remove array functions from getPassByReferenceFunctions
The list of pass-by-reference functions is only currently used to
determine if a function's argument might be counted as a define. All
arguments passed to the function are counted as a read. Functions like
`array_pop()` will throw a PHP warning if used with an undefined
variable, so this removes them. There are probably many others in this
list that can be removed, but this is a start.