Mark concatenation inside another expression as variable read (#186)
* Add test case for concatination on the same line as assignment

* Remove array functions from getPassByReferenceFunctions

The list of pass-by-reference functions is only currently used to
determine if a function's argument might be counted as a define. All
arguments passed to the function are counted as a read. Functions like
`array_pop()` will throw a PHP warning if used with an undefined
variable, so this removes them. There are probably many others in this
list that can be removed, but this is a start.

* Add variations of concat-and-assign to test fixture

* Add function expression to assign-and-concat test

* Mark assignments as read if they are part of another expression

* Fix test for undefined variable after array assignment in loop

* Just check for RHS

* Add check for inside function calls

* Replace bool casts for indices with is_int

Indices can sometimes be 0.