Composer.json - widen PHPUnit requirements
PHP 5.4 needs PHPUnit 4.x to be able to run the tests, so that needed adding.
Aside from that PHPUnit 9.x has come out quite a while back and has better support for PHP 8.0 and 8.1 than PHPUnit 8.x.

I've set the version constraints for the 4.x and 5.x series to very specific minimum versions for a reason: those are the first versions to have the forward-compatibility layer with the namespaced PHPUnit `TestCase`, which is used by the tests in this repo.