GH Actions: fix CI
Grrr....

When PHPUnit has been installed on a high PHP version, some of the dependencies of PHPUnit will now be installed in versions not compatible with PHP 7.4, which blocks the install of the Coveralls package.

Installing PHP Coveralls globally instead should fix it.

I just wish PHP Coveralls would finally release a version compatible with PHP >  8.0....